style(mesh_sampling): move template implementation to .hpp
